+++ aws/middleware/middleware_test.go
@@ -16,14 +16,14 @@ import (
 	smithyhttp "github.com/awslabs/smithy-go/transport/http"
 )
 
-func TestRequestInvocationIDMiddleware(t *testing.T) {
+func TestClientRequestID(t *testing.T) {
 	oReader := rand.Reader
 	defer func() {
 		rand.Reader = oReader
 	}()
 	rand.Reader = bytes.NewReader(make([]byte, 16))
 
-	mid := middleware.RequestInvocationIDMiddleware{}
+	mid := middleware.ClientRequestID{}
 
 	in := smithymiddleware.BuildInput{Request: &smithyhttp.Request{Request: &http.Request{Header: make(http.Header)}}}
 	ctx := context.Background()