+++ service/internal/s3shared/response_error_middleware.go
@@ -4,16 +4,17 @@ import (
 	"context"
 
 	awsmiddleware "github.com/aws/aws-sdk-go-v2/aws/middleware"
+	"github.com/aws/aws-sdk-go-v2/aws/middleware/id"
 	awshttp "github.com/aws/aws-sdk-go-v2/aws/transport/http"
 	"github.com/awslabs/smithy-go/middleware"
 	smithyhttp "github.com/awslabs/smithy-go/transport/http"
 )
 
 // AddResponseErrorMiddleware adds response error wrapper middleware
-func AddResponseErrorMiddleware(stack *middleware.Stack) {
+func AddResponseErrorMiddleware(stack *middleware.Stack) error {
 	// add error wrapper middleware before request id retriever middleware so that it can wrap the error response
 	// returned by operation deserializers
-	stack.Deserialize.Insert(&errorWrapperMiddleware{}, "S3MetadataRetrieverMiddleware", middleware.Before)
+	return stack.Deserialize.Add(&errorWrapperMiddleware{}, middleware.Before)
 }
 
 type errorWrapperMiddleware struct {
@@ -21,7 +22,7 @@ type errorWrapperMiddleware struct {
 
 // ID returns the middleware identifier
 func (m *errorWrapperMiddleware) ID() string {
-	return "S3ResponseErrorWrapperMiddleware"
+	return id.ResponseErrorWrapper
 }
 
 func (m *errorWrapperMiddleware) HandleDeserialize(ctx context.Context, in middleware.DeserializeInput, next middleware.DeserializeHandler) (